Change calling instead of var move to let; Move stackview
